+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Released ...
 
 Summary: TODO
 
+- Python: added `__geo_interface__` to mapnik.Feature and mapnik.Path (#2009)
+
 - Python: Exposed optimized WKTReader for parsing WKT into geometry paths (6bfbb53)
 
 - Optimized expression evaluation of text by avoiding extra copy (1dd1275)